+++ src/CollectionInterface.php
@@ -4,16 +4,36 @@ declare(strict_types=1);
 
 namespace Steevanb\PhpCollection;
 
-use Steevanb\PhpCollection\{
-    ScalarCollection\IntegerCollection,
-    ScalarCollection\StringCollection
-};
-
-/** @template T */
+/** @template TValueType */
 interface CollectionInterface extends \Countable
 {
-    /** @param iterable<T> $values */
-    public function __construct(iterable $values = []);
+    /** @param TValueType $value */
+    public function set(string|int $key, mixed $value): static;
+
+    /** @param iterable<string|int, TValueType> $values */
+    public function replace(iterable $values): static;
+
+    /** @param TValueType $value */
+    public function add(mixed $value): static;
+
+    /** @return TValueType */
+    public function get(string|int $key): mixed;
+
+    public function remove(string|int $key): static;
+
+    public function clear(): static;
+
+    public function isEmpty(): bool;
+
+    /** @param TValueType $value */
+    public function contains(mixed $value): bool;
+
+    public function resetKeys(): static;
+
+    /** @return array<int, string|int> */
+    public function getKeys(): array;
+
+    public function hasKey(string|int $key): bool;
 
     public function setReadOnly(bool $readOnly): static;
 
@@ -21,30 +41,6 @@ interface CollectionInterface extends \Countable
 
     public function getValueAlreadyExistsMode(): ValueAlreadyExistsModeEnum;
 
-    public function hasKey(string|int $key): bool;
-
-    /** @return array<int, string|int> */
-    public function getKeys(): array;
-
-    public function getIntegerKeys(): IntegerCollection;
-
-    public function getStringKeys(): StringCollection;
-
-    public function resetKeys(): static;
-
-    /** @return T */
-    public function get(string|int $key): mixed;
-
-    /** @param T $value */
-    public function contains(mixed $value): bool;
-
-    public function remove(string|int $key): static;
-
-    /** @param iterable<T> $values */
-    public function replace(iterable $values): static;
-
-    public function clear(): static;
-
-    /** @return array<string|int, T> */
+    /** @return array<string|int, TValueType> */
     public function toArray(): array;
 }