+++ src/ScalarCollection/FloatNullableCollection.php
@@ -12,21 +12,6 @@ use Steevanb\PhpCollection\{
 /** @extends AbstractCollection<float|null> */
 class FloatNullableCollection extends AbstractCollection
 {
-    public function set(int|string $key, float|null $value): static
-    {
-        return $this->doSet($key, $value);
-    }
-
-    public function add(float|null $value): static
-    {
-        return $this->doAdd($value);
-    }
-
-    public function merge(FloatCollection|FloatNullableCollection $collection): static
-    {
-        return $this->doMerge($collection);
-    }
-
     protected function canAddValue(mixed $value): bool
     {
         if (is_null($value) === false && is_float($value) === false) {